comments: save comments in database

+++ server/app/Controllers/Http/AdminOriginsController.ts
@@ -1,6 +1,6 @@
 import { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
 
-import Origin from 'App/Models/Origin'
+import Origin, { OriginTypes } from 'App/Models/Origin'
 import OriginValidator from 'App/Validators/OriginValidator'
 import OriginYouTubeValidator from 'App/Validators/OriginYouTubeValidator'
 import YoutubeProvider from '@ioc:Providers/YouTube'
@@ -13,6 +13,10 @@ export default class OriginsController {
 
   public async store({ request }: HttpContextContract) {
     const { name, type, config } = await request.validate(OriginValidator)
+
+    if (type === OriginTypes.Main) {
+      throw new Error("can't create main origin")
+    }
 
     if (type === 'you-tube') {
       await request.validate(OriginYouTubeValidator)
@@ -34,7 +38,9 @@ export default class OriginsController {
 
   public async destroy({ params }: HttpContextContract) {
     const origin = await Origin.findOrFail(params.id)
-    await origin.related('videos').query().delete()
+    if (origin.type === OriginTypes.Main) {
+      throw new Error("can't delete main origin")
+    }
     await origin.delete()
   }
 }
